+++ jkorm/src/main/kotlin/com/tang/jkorm/session/SqlSession.kt
@@ -41,7 +41,7 @@ interface SqlSession : AutoCloseable {
 
     fun <T> count(method: Method, mapperInterface: Class<T>, type: Class<T>, parameter: Any?): Long
 
-    fun <T> paginate(method: Method, mapperInterface: Class<T>, type: Class<T>, pageNumber: Long, pageSize: Long, parameter: Any?, orderBys: Array<OrderItem>): Page<T>
+    fun <T> paginate(method: Method, mapperInterface: Class<T>, type: Class<T>, pageNumber: Long, pageSize: Long, parameter: Any?, orderBys: Array<OrderItem<T>>): Page<T>
 
     fun commit()