+++ CHANGELOG.md
@@ -28,3 +28,4 @@
  * Support for configuration using spring boot
  * Add `Pairs` class to simplify `Pair` operation in Kotlin
  * Add `OrderItem` for paginate order by
+ * Add `OrderItem` for paginate order by with generics